[RateLimiter] fix incorrect retryAfter of FixedWindow
A FixedWindow always ends `intervalInSeconds` after the start time
(`timer`). So when calculating the time to consume some tokens the
tokens will always be available at `timer + intervalInSeconds`. But
currently this is reported incorrectly as `calculateTimeForTokens()`
calculates the time based on the desired amount of tokens (and cycles)
while it doesn't take into account `maxSize` amount of tokens become
available at the windows end.

Furthermore calculating the amount of needed cycles is incorrect. This
as all tokens become available at once (at the windows end) and you
can't consume more tokens than `maxSize` (which is validated at the
start of `FixedWindowLimiter::reserve`, in case of `tokens > limit` it
throws).